Feature&Fix/GameSystem: 천장 기울어짐 이상현상 추가, AnomalyID 버그 수정

+++ Source/Endless_Hotel/Anomaly/Object/Anomaly_Object_Base.h
@@ -22,7 +22,7 @@ public:
 
 public:
 	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = "Object|ID")
-	int32 ObjectID = -1;
+	uint8 ObjectID = -1;
 
 #pragma endregion
